Move SnakeDef into a new file
